separate component for product list
